# 3x speedup of `multiplier_bootstrap`