@bcallaway11 2023-08-28
Hey Kyle, this looks cool, I was looking at the code.  Do you think that we should have any worries about compatibility issues across systems (based on the number of bits part) or extremely large sample sizes?