@kylebutts 2023-08-29
Hi @bcallaway11,

Re bits: R defines integers as 32 bits on all systems. Since I'm drawing random ints between 0 and 2^31 - 1 = 2147483647, each bit of the integer equal 0 and 1 with probability 1/2. However, RNG is usually the slow step, so this is a big speedup as you generate 31 random 0/1 variables per RNG draw. In this step, memory scales with the number of rows of `inf_func`

Re memory: I don't think it should create problems with really big datasets. Memory usage scales at (biters x K) with K being the number of columns of the influence function matrix. I'm not sure why it's using more memory in the second reprex. Will look into this ...