@kylebutts 2023-08-29
Okay, profiled and basically 1/2 the memory usage is just converting a R matrix to an `arma::mat` which is unavoidable (to use RcppArmadillo's speedups). The other 1/2 is from RNG which uses a lot of memory but not high peak memory anymore (after my latest commit).

Memory:
![Memory Usage bootstrap_multiplier](https://github.com/bcallaway11/BMisc/assets/19961439/5361d344-b73b-405c-b5d8-eb0d2eaa0e8f)

Timewise, almost all the time now is doing the dot-product of columns with the vector. It calls `_ddot` in BLAS, so that's highly optimized. Not sure there's much speed-up left in this function :-)

Time:
![Time Usage bootstrap_multiplier](https://github.com/bcallaway11/BMisc/assets/19961439/5059e405-77ab-48b4-9422-fb72b7b17336)