@bcallaway11 2023-08-30
FYI, I just ran all our unit tests in `did`, I wasn't checking carefully, but there weren't any issues, and about 20% improvement in computation speed (not all of these tests would be affected by the updates here either), so this seems to work great.  Thanks a lot.